+++ src/complex.hpp
@@ -42,7 +42,7 @@ public:
 
     int16_t GetId() const;
 
-    uint16_t GetTypeIndex() const;
+    uint32_t GetTypeIndex() const;
     void FileLoad(SmartFileReader& file) noexcept;
     void FileSave(SmartFileWriter& file) noexcept;
